Split Name::Change tests out of name_test.rb (#4708)
Moves the 5 tests exercising Name::Change (is_misspelling?/correct_spelling
side effects of change_deprecated, change_author, change_text_name's
parent-creation-failure guard, mark_misspelled, clear_misspelled) into
test/models/name/change_test.rb, matching app/models/name/change.rb.
Includes a local copy of create_test_name (used by
test_changing_author_of_autonym).

Also fixes a bug from the earlier Taxonomy stragglers commit: the
change_text_name doc comment was swept into taxonomy_test.rb by an
off-by-N range (test_changing_classification_propagates_to_subtaxa's
real closing `end` was 6 lines earlier than assumed). Removed the
orphaned comment from taxonomy_test.rb and restored it above
test_change_text_name_raises_when_parent_creation_fails here.

require_relative'd from name_test.rb so running that file alone still
runs the full original suite.
